+++ src/app/admin/components/seasons-list/seasons-list.component.ts
@@ -8,6 +8,7 @@ import { SeasonsService } from "../../services/admin-seasons.service";
 import { AdminSeasonsCreateModal } from "../seasons-create/seasons-create.component";
 
 import { globalFunctions } from "../../../functions/global.function";
+import { AdminConfirmationDialogModal } from "../confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: 'app-admin-seasons-list',
@@ -27,8 +28,8 @@ export class AdminSeasonsList implements OnInit, OnDestroy {
                               ];
 
   constructor(
-    public seasonsService: SeasonsService,
-    public dialog: MatDialog,
+    private seasonsService: SeasonsService,
+    private dialog: MatDialog,
     private globalFunctions: globalFunctions
   ) {}
 
@@ -61,9 +62,21 @@ export class AdminSeasonsList implements OnInit, OnDestroy {
   }
 
   onDelete(id: number) {
-    this.isLoading = true;
-    this.seasonsService.deleteSeason(id);
-    this.isLoading = false;
+    const dialogRef = this.dialog.open(AdminConfirmationDialogModal, {
+      data: {
+        title: 'İŞLEMİ ONAYLIYOR MUSUNUZ?',
+        message: 'Bu işlem verilerinizi kalıcı olarak silecektir, işleminizi onaylıyor musunuz?'
+      }
+    });
+
+    dialogRef.afterClosed()
+      .subscribe({
+        next: (data) => {
+          if (data) {
+            this.seasonsService.deleteSeason(id);
+          }
+        }
+      });
   }
 
   ngOnDestroy() {
